+++ lib/dbms.js
@@ -1,12 +1,25 @@
 'use strict';
 
 const path = require('path');
-const fs = require('fs').promises;
+const fsp = require('fs').promises;
 const { Model } = require('metaschema');
 
-const { cpdir, shorten } = require('./utils.js');
-
 const dbms = {};
+
+const cpdir = async (from, to) => {
+  const files = await fsp.readdir(from, { withFileTypes: true });
+  for (const file of files) {
+    if (file.isDirectory()) continue;
+    const fromPath = path.join(from, file.name);
+    const toPath = path.join(to, file.name);
+    await fsp.copyFile(fromPath, toPath);
+  }
+};
+
+const shorten = (dir) => {
+  const pos = dir.lastIndexOf('/');
+  return pos !== -1 ? dir.substring(pos) : dir;
+};
 
 const loadModel = async (modelPath) => {
   const dbmsTypes = dbms.pg.types;
@@ -19,7 +32,7 @@ const loadModel = async (modelPath) => {
 
 const getPreviousVersion = async (modelPath) => {
   const historyPath = path.join(modelPath, 'history');
-  const folders = await fs.readdir(historyPath, { withFileTypes: true });
+  const folders = await fsp.readdir(historyPath, { withFileTypes: true });
   let version = 0;
   let previousName = '';
   for (const folder of folders) {
@@ -45,7 +58,7 @@ const create = async (modelPath, outputPath) => {
     script.push(sql);
   }
   const dbPath = path.join(outputPath, 'database.sql');
-  await fs.writeFile(dbPath, script.join('\n\n') + '\n');
+  await fsp.writeFile(dbPath, script.join('\n\n') + '\n');
   const dtsPath = path.join(outputPath, 'database.d.ts');
   await model.saveTypes(dtsPath);
 };
@@ -73,7 +86,7 @@ const generate = async (modelPath) => {
   const migUp = mig + '-up.sql';
   const migDn = mig + '-dn.sql';
   console.log(`Save history: ${shorten(newFolder)}`);
-  await fs.mkdir(newFolder);
+  await fsp.mkdir(newFolder);
   await cpdir(modelPath, newFolder);
   console.log(`Migration up: ${shorten(migUp)}`);
   console.log(`Migration down: ${shorten(migDn)}`);
@@ -85,7 +98,7 @@ const migrate = async (modelPath, version) => {
   if (version) console.log(`Migration to this version: ${version}`);
   else console.log('Migration to the latest version');
   const migPath = path.join(modelPath, `migrations`);
-  const files = await fs.readdir(migPath, { withFileTypes: true });
+  const files = await fsp.readdir(migPath, { withFileTypes: true });
   files.sort((a, b) => (a.name < b.name ? -1 : 1));
   for (const file of files) {
     if (file.isDirectory()) continue;
@@ -96,7 +109,7 @@ const migrate = async (modelPath, version) => {
     if (!name.endsWith('-up.sql')) continue;
     if (!version || v <= version) {
       const fileName = path.join(migPath, name);
-      const sql = await fs.readFile(fileName, 'utf8');
+      const sql = await fsp.readFile(fileName, 'utf8');
       console.log(`Apply script: ${name}`);
       await execute(sql);
     }
