+++ lib/pg.js
@@ -2,10 +2,10 @@
 
 const { Client } = require('pg');
 
-const { dbms } = require('./model-db.js');
+const { isFirstUpper, toLowerCamel, toUpperCamel } = require('metautil');
 const { Schema } = require('metaschema');
 
-const { isFirstUpper, toLowerCamel, toUpperCamel } = require('metautil');
+const { dbms } = require('./dbms.js');
 
 const DB_RELATION = 1;
 const DB_FIELD = 2;
